rename users service and controller to auth
