+++ src/routes/authRouter.ts
@@ -1,7 +1,7 @@
 import dotenv from "dotenv";
 import express from "express";
 import passport from "passport";
-import userController from "../controllers/userController";
+import userController from "../controllers/authController";
 import { validationMiddleware } from "../middleware/validationMiddleware";
 import {
   LoginValidationSchema,
