+++ src/services/index.ts
@@ -1,3 +1,3 @@
 export { default as productService } from "./productsService";
 export { default as orderService } from "./ordersService";
-export { default as userService } from "./usersService";
+export { default as userService } from "./authService";
